Changed checks to return an array of [status, message]
This change allows us to simplify checks greatly.  Now the check
doesn’t need to know it’s name or whether it’s critical or not, it just
needs to know how to run, and return it’s status and the message.

This allows us to get rid of the base Check object all together.